iter 2/requirements.txt
mysql-connector-python==8.0.27
pandas==1.3.3
matplotlib==3.4.3